Do not format ruby code in erb-formatter
`erb-formatter` uses the `syntax_tree` gem which has its own
formatting ideas about ruby, but, in our code base, so does Rubocop,
the latter which we configure extensively.  When they disagree, the
result is flip-flopping between tools, and that can fail the clover
build, in this CI step in `.github/workflows/ci.yml`:

    - name: Check that source code formatters and annotations make no changes
      run: git diff --stat --exit-code

This was motivated by an attempt to this `.rubocop.yml` configuration
in clover:

    Layout/DotPosition:
      EnforcedStyle: trailing

In this case, Rubocop moves method-chained periods to the ends of
lines, and then `erb-formatter` moves them back to the beginning
again.  This fails CI.

Given our more extensive rubocop use, and use of the `rubocop-erb` as
we also use it for rewriting, it's simpler to remove `syntax_tree` as
a dependency altogether, and not attempt to format code in
`erb-formatter`.  This keeps `erb-formatter` the rest of its
functionality, such as tag indentation.